requirements.txt
openai
tiktoken
python-dotenv
requests
mcp
httpx
uvicorn
fastapi
groq
duckduckgo-search
pydantic
